fix: improve narrow implementation
